Adjust test suite for better skipped test output
runtests now always prints out the count of skipped tests even if there
are failures or aborted tests.  Adjust the tests accordingly.  Also drop
the skip test from the abort test set so that there will be only one
skipped test, to add a test of the pluralization of the skipped count.