Reorganize the TAP support library
libtest is now libtap, more accurately reflecting what it's for.  Move
the source into a tap subdirectory and move libtest.h to tap/basic.h to
make it easier for other packages to add their own additional code to
the libtap library.  Change the tests as appropriate for the renaming
and catch one t to tests renaming missing in libtap.t.