Add knowledge of source and build directories to runtests
Set the SOURCE and BUILD environment variables when running tests to
the roots of the source and build test directories if set at compile
time or via the -s and -b command-line options to runtests.

Search for test programs first in the current directory and then in
the build and source directories, in that order, as specified by
SOURCE and BUILD or by the -s and -b options.  This allows some tests
to be compiled and some shipped with the package and merged together
at run time by runtests, even if the build directory is not the source
directory.

Also remove the unnecessary internal function prototypes in runtests
and restructure how it searches for test programs so that it's
encapsulated in its own function.