Take advantage of the test suite BUILD/SOURCE support
Take advantage of the new BUILD/SOURCE support to have considerably fewer
Autoconf-processed files.

Also change the banner of runtests to always omit the path to the tests
file, since otherwise the output is not reproducible in the test suite
for runtests itself.

Remove the accidentally-committed tests/env/env.t, which is generated by
Autoconf still, and update .gitignore properly.