Allow omitting the test number in ok and not ok lines
Add support for omitting the test number from an "ok" or "not ok"
line, in which case the number is assumed to be one more than the
previous number.

Omitting the space after "ok" or "not ok" line is now supported, and a
test number of 0 now results in an error.