Add support for specifying the plan anywhere in the test output
Support giving the plan at any point in the test output rather than
requiring it be first.  The TAP protocol permits giving the plan last
instead of first, but since it's just as easy to support providing the
plan in the middle of the test output, allow that as well.