Simplify and slightly improve output formatting for failing summaries
Simplify and improve output formatting in the summary of failing tests
in some edge cases.  Don't miscount the output length if printf fails
for some reason.  Use a limit for the test summary that matches the
length of the header lines.

Fix a few other signed/unsigned integer conversion issues.