Major source reorganization
Reorganize the source package so that runtests.c is in the tests
directory and the C and shell TAP libraries are in tests/tap.  Change
the include directives in the C TAP library source files accordingly.
This is somewhat confusing for the standalone package, but means that
the files are located in the same locations here as they would be in
packages into which they're copied, which simplifies instructions and
the build system.  This also removes the need to build the C TAP
library with -I flags at the cost of requiring that it be located in a
tests subdirectory, allowing easier use of additional supporting
headers without requiring that all tests be built with a -I directive.