Document decision to ignore TAP version directives
These are currently treated as unknown lines and ignored.  This seems
to be correct for now; the current TAP output format doesn't use them,
and the right approach for any future version of TAP would seem to be
to try to parse it anyway.  That ends up being equivalent to ignoring
the version line.