Check integer overflows on memory allocation, fix string creation
Check for integer overflows on memory allocation.  All the possible
issues for this code are rather theoretical, but one may as well
strive for correctness.

Replace all uses of sprintf with a simpler string concatenation
function that checks for allocation overflow.  (The standards
assumptions for this package don't permit assuming asprintf or a
sufficiently non-broken snprintf to simulate asprintf.)