Various improvements to verbose support
Add a -v command-line option that does the same thing, and rename
the environment variable to C_TAP_VERBOSE.  Clean up the output
considerably.  It's still a bit basic, but it's not quite as ugly,
although that required sprinkling lots more conditionals around.

Add a test suite and update the documentation.