fix: Remove dead code (Revert "Fix x_reallocarray guard")
This reverts commit 599c10937af8a5c1b6f779df33c13a363de61dd0.

The reverted patch introduced again a dead code check.
The variable n can never 0 in this location and a divide by zero is never possible

Long code flow description:
For the case, that the function x_reallocarray is called with zero for n,
the first line in the function x_reallocarray makes sure, that n is at least 1
for the following code.
```
n = (n > 0) ? n : 1;
```
https://github.com/rra/c-tap-harness/blob/main/tests/runtests.c#L408

The reverted commit added a check "n > 0", which can never fail.

The code was fixed in Sep. 2021, but the reverted commit added the dead code again.

--
Regards ... Detlef